Спринт 15 исправления по рекомендациям
