+++ app/src/main/java/com/practicum/playlistmaker/data/repository/SearchRepositoryImpl.kt
@@ -1,9 +1,9 @@
 package com.practicum.playlistmaker.data.repository
 
-import com.practicum.playlistmaker.data.dto.TrackDto
 import com.practicum.playlistmaker.data.network.ItunesApiService
 import com.practicum.playlistmaker.domain.api.SearchRepository
 import com.practicum.playlistmaker.domain.models.Track
+import com.practicum.playlistmaker.data.toTrack
 
 class SearchRepositoryImpl(
     private val apiService: ItunesApiService
@@ -12,12 +12,13 @@ class SearchRepositoryImpl(
     override fun searchTracks(query: String): List<Track> {
         val response = apiService.search(query).execute()
         return if (response.isSuccessful && response.body() != null) {
-            response.body()!!.results.map { trackDto -> trackDto.toTrack() }
+            response.body()!!.results.map { it.toTrack() } //{ trackDto -> trackDto.toTrack() }
         } else {
             emptyList()
         }
     }
 
+    /*
     private fun TrackDto.toTrack(): Track {
         return Track(
             trackId = trackId,
@@ -31,5 +32,5 @@ class SearchRepositoryImpl(
             country = country,
             previewUrl = previewUrl
         )
-    }
+    }*/
 }
